Update app-centric-management.md
Added text to consider using Set-UnifiedGroup PowerShell to disable the welcome message for adding members to the Unified Group which is not present in the UI and can only be done by PowerShell.  Without doing this in a large enterprise environment this can generate a lot of emails to users.